refactor: rename pkg name
