+++ packages/mask/package.json
@@ -43,7 +43,7 @@
     "@masknet/shared": "workspace:*",
     "@masknet/shared-base": "workspace:*",
     "@masknet/shared-base-ui": "workspace:*",
-    "@masknet/social-network-infra": "workspace:*",
+    "@masknet/types": "workspace:*",
     "@masknet/theme": "workspace:*",
     "@masknet/typed-message": "workspace:*",
     "@masknet/web3-constants": "workspace:*",
