+++ packages/mask/src/social-network-adaptor/browser-action/index.ts
@@ -1,4 +1,4 @@
-import type { SocialNetwork, SocialNetworkUI } from '@masknet/social-network-infra'
+import type { SocialNetwork, SocialNetworkUI } from '@masknet/types'
 import { defineSocialNetworkUI, definedSocialNetworkUIs } from '../../social-network/index.js'
 import { isEnvironment, Environment, ValueRef } from '@dimensiondev/holoflows-kit'
 import { SocialNetworkEnum } from '@masknet/encryption'
