+++ packages/mask/src/social-network-adaptor/facebook.com/base.ts
@@ -1,6 +1,6 @@
 import { SocialNetworkEnum } from '@masknet/encryption'
 import { EnhanceableSite } from '@masknet/shared-base'
-import type { SocialNetwork } from '@masknet/social-network-infra'
+import type { SocialNetwork } from '@masknet/types'
 
 const origins = ['https://www.facebook.com/*', 'https://m.facebook.com/*', 'https://facebook.com/*']
