+++ packages/mask/src/social-network-adaptor/instagram.com/collecting/identity.ts
@@ -1,7 +1,7 @@
 import { LiveSelector, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
 import { delay } from '@masknet/kit'
 import { ProfileIdentifier } from '@masknet/shared-base'
-import { SocialNetworkUI as Next, CREATOR } from '@masknet/social-network-infra'
+import { SocialNetworkUI as Next, CREATOR } from '@masknet/types'
 import { instagramBase } from '../base.js'
 import { searchInstagramAvatarSelector } from '../utils/selector.js'
 import { getAvatar, getBioDescription, getNickname, getPersonalHomepage, getUserId } from '../utils/user.js'
