+++ packages/mask/src/social-network-adaptor/mirror.xyz/collecting/posts.ts
@@ -1,5 +1,5 @@
 import { DOMProxy, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
-import { SocialNetworkUI as Next, CREATOR } from '@masknet/social-network-infra'
+import { SocialNetworkUI as Next, CREATOR } from '@masknet/types'
 import { postsContentSelector } from '../utils/selectors.js'
 
 import { mirrorShared } from '../shared.js'
