+++ packages/mask/src/social-network-adaptor/opensea.io/shared.ts
@@ -1,4 +1,4 @@
-import type { SocialNetwork } from '@masknet/social-network-infra'
+import type { SocialNetwork } from '@masknet/types'
 import { createSNSAdaptorSpecializedPostContext } from '../../social-network/utils/create-post-context.js'
 import { hasPayloadLike } from '../../utils/index.js'
 import { openseaBase } from './base.js'
