+++ packages/mask/src/social-network-adaptor/opensea.io/ui-provider.ts
@@ -1,4 +1,4 @@
-import { SocialNetworkUI, STATE_CREATOR } from '@masknet/social-network-infra'
+import { SocialNetworkUI, STATE_CREATOR } from '@masknet/types'
 import { createTaskStartSetupGuideDefault } from '../../social-network/defaults/index.js'
 import { InitAutonomousStateProfiles } from '../../social-network/defaults/state/InitProfiles.js'
 import { IdentityProviderOpensea } from './collecting/Identity.js'
