+++ packages/mask/src/social-network-adaptor/twitter.com/automation/pasteTextToComposition.ts
@@ -1,7 +1,7 @@
 import { delay } from '@masknet/kit'
 import { inputText, pasteText } from '@masknet/injected-script'
 import { newPostButtonSelector, postEditorDraftContentSelector } from '../utils/selector.js'
-import type { SocialNetworkUI } from '@masknet/social-network-infra'
+import type { SocialNetworkUI } from '@masknet/types'
 import { getEditorContent, hasEditor, hasFocus, isCompose } from '../utils/postBox.js'
 import { untilElementAvailable } from '../../../utils/dom.js'
 import { isMobileTwitter } from '../utils/isMobile.js'
