+++ packages/mask/src/social-network-adaptor/twitter.com/collecting/post.ts
@@ -2,7 +2,7 @@ import { memoize, noop } from 'lodash-unified'
 import { PostIdentifier, ProfileIdentifier } from '@masknet/shared-base'
 import { IntervalWatcher, DOMProxy, DOMProxyEvents } from '@dimensiondev/holoflows-kit'
 import type { EventListener } from '@servie/events'
-import { SocialNetworkUI as Next, CREATOR } from '@masknet/social-network-infra'
+import { SocialNetworkUI as Next, CREATOR } from '@masknet/types'
 import type { PostInfo } from '@masknet/plugin-infra/content-script'
 import { globalUIState } from '../../../social-network/index.js'
 import { postIdParser, postParser, postImagesParser, postContentMessageParser } from '../utils/fetch.js'
