+++ packages/mask/src/social-network/defaults/automation/AttachImageToComposition.ts
@@ -1,6 +1,6 @@
 import { downloadUrl, MaskMessages, pasteImageToActiveElements } from '../../../utils/index.js'
 import { delay, waitDocumentReadyState } from '@masknet/kit'
-import type { SocialNetworkUI } from '@masknet/social-network-infra'
+import type { SocialNetworkUI } from '@masknet/types'
 import { activatedSocialNetworkUI } from '../../ui.js'
 
 export function pasteImageToCompositionDefault(hasSucceed: () => Promise<boolean> | boolean) {
