+++ packages/mask/src/social-network/defaults/state/InitProfiles.ts
@@ -1,6 +1,6 @@
 import { MaskMessages } from '../../../utils/messages.js'
 import Services from '../../../extension/service.js'
-import type { SocialNetworkUI } from '@masknet/social-network-infra'
+import type { SocialNetworkUI } from '@masknet/types'
 import type { ValueRef } from '@dimensiondev/holoflows-kit'
 import type { ProfileInformation } from '@masknet/shared-base'
