+++ packages/types/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@masknet/social-network-infra",
+  "name": "@masknet/types",
   "private": true,
   "exports": {
     ".": {
