+++ tsconfig.json
@@ -91,7 +91,7 @@
       "@masknet/web3-shared-solana": ["./packages/web3-shared/solana/src"],
       "@masknet/typed-message": ["./packages/typed-message/base"],
       "@masknet/typed-message/dom": ["./packages/typed-message/dom"],
-      "@masknet/social-network-infra": ["./packages/social-network-infra/src"],
+      "@masknet/types": ["./packages/types/src"],
       "@masknet/sandboxed-plugin-runtime": ["./packages/sandboxed-plugin-runtime/src/runtime/export.ts"],
       "@masknet/sandboxed-plugin-runtime/background": ["./packages/sandboxed-plugin-runtime/src/background"],
       "@masknet/sandboxed-plugin-runtime/site-adaptor": ["./packages/sandboxed-plugin-runtime/src/site-adaptor"],
